+++ examples/visualization.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-# (c) 2018 Open Risk, all rights reserved
+# (c) 2018-2019 Open Risk, all rights reserved
 #
 # DataQualityToolkit is licensed under the Apache 2.0 license a copy of which is included
 # in the source distribution of TransitionMatrix. This is notwithstanding any licenses of
@@ -20,20 +20,21 @@ import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.cm as cm
 from DQToolkit import Rule
-from DQToolkit import DataSource
+from DQToolkit import WWWDataSource
 
-filename = "../datasets/EBA_Sample.xlsx"
-header_row = 2
+# Set the remote URL where we will fetch data from (obviously it must be accessible!)
+url = "https://en.wikipedia.org/wiki/List_of_data_breaches"
 
 # Instantiate a validation rule object
 MyRule = Rule()
+# Check the available rules
+MyRule.rule_data()
+
 MyRule.activate('R4')
 
 # Instantiate a datasource object
-MySource = DataSource(filename, header_row)
+MySource = WWWDataSource(url)
 MySource.validate_all(MyRule)
-# MySource.visualize()
-
 r = []
 theta = []
 colors = []
@@ -41,19 +42,19 @@ dt = 2 * np.pi / MySource.result_no
 dr = 0.02
 ts = 0
 
-for frame in MySource.results:
-    for col in MySource.results[frame]:
-        ts = ts + dt
-        rs = 0.1
-        if MySource.results[frame][col] is not None:
-            for point in MySource.results[frame][col]:
-                rs = rs + dr
-                r.append(rs)
-                theta.append(ts)
-                if point == np.True_:
-                    colors.append('azure')
-                else:
-                    colors.append('greenyellow')
+
+for col in MySource.results:
+    ts = ts + dt
+    rs = 0.1
+    if MySource.results[col] is not None:
+        for point in MySource.results[col]:
+            rs = rs + dr
+            r.append(rs)
+            theta.append(ts)
+            if point == np.True_:
+                colors.append('azure')
+            else:
+                colors.append('greenyellow')
 
 fig = plt.figure(facecolor='#0B0050')
 fig.suptitle('OpenCPM::DQToolkit', fontsize=20, color='azure')