+++ CHANGELOG.rst
@@ -1,7 +1,19 @@
 DataQualityToolkit Change Log
-===========================
+=============================
 
-v0.1.0 (22-06-2019)
+v0.2.1 (09-01-2019)
+===================
+* First extension to capture web data sources
+
+
+v0.2.0 (09-01-2019)
+===================
+
+* Added sphinx documentation
+* Self documenting Rule object
+
+
+v0.1.0 (22-06-2018)
 ===================
 
 * First release of the package